requirements.txt
Django==5.2.2
djangorestframework==3.15.2
django-ckeditor==6.7.1
django-cors-headers==4.3.1
beautifulsoup4==4.12.2
pytz==2024.1
gunicorn==23.0.0
Pillow==11.3.0
requests==2.32.4
openai==1.51.2
python-decouple==3.8 